@@ -1,15 +1,14 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
+    <TargetFramework>net10.0</TargetFramework>
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
 
   <ItemGroup>
-	  <PackageReference Include="Swashbuckle.AspNetCore" Version="8.1.4" />
-	  <PackageReference Include="Microsoft.OpenApi" Version="1.6.28" />
-	  <PackageReference Include="Swashbuckle.AspNetCore.SwaggerGen" Version="8.1.4" />
+	  <PackageReference Include="Swashbuckle.AspNetCore" Version="10.1.0" />
+	  <PackageReference Include="Swashbuckle.AspNetCore.SwaggerGen" Version="10.1.0" />
   </ItemGroup>
 
 </Project>
