@@ -9,13 +9,13 @@ public class KoalesceOptions : IValidatableObject
 	/// <summary>
 	/// Source URLs. At least one source is required.
 	/// </summary>
-	[Required]
+	[Required]	
 	public List<ApiSource> Sources { get; set; } = new();
 
 	/// <summary>
 	/// The logical path where the merged definition should be exposed.
-	/// </summary>
-	[Required]
+	/// </summary>	
+	[Required]	
 	public string MergedDocumentPath { get; set; } = default!;
 
 	/// <summary>
@@ -50,6 +50,12 @@ public class KoalesceOptions : IValidatableObject
 	public bool FailOnServiceLoadError { get; set; } = false;
 
 	/// <summary>
+	/// HTTP request timeout in seconds for fetching API specifications.
+	/// Default: 15 seconds.
+	/// </summary>
+	public int HttpTimeoutSeconds { get; set; } = CoreConstants.DefaultHttpTimeoutSeconds;
+
+	/// <summary>
 	/// Custom validation logic for required fields.
 	/// </summary>	
 	public virtual IEnumerable<ValidationResult> Validate(ValidationContext validationContext)
@@ -62,9 +68,9 @@ public class KoalesceOptions : IValidatableObject
 				[nameof(Sources)]);
 		}
 		else
-		{			
-			foreach (var validationResult in ValidateSourceUrls()) 
-				yield return validationResult;			
+		{
+			foreach (var validationResult in ValidateSources())
+				yield return validationResult;
 		}
 
 		// Validating MergedDocumentPath
@@ -124,10 +130,32 @@ public class KoalesceOptions : IValidatableObject
 					[nameof(SchemaConflictPattern)]);
 			}
 		}
+
+		// Validate HttpTimeoutSeconds
+		if (HttpTimeoutSeconds <= 0)
+		{
+			yield return new ValidationResult(
+				CoreConstants.HttpTimeoutMustBePositive,
+				[nameof(HttpTimeoutSeconds)]);
+		}
 	}
 
-	private IEnumerable<ValidationResult> ValidateSourceUrls()
+	private IEnumerable<ValidationResult> ValidateSources()
 	{
+		// Check for duplicate sources (same Url or FilePath)
+		var sourceGroups = Sources
+			.Select((source, index) => (source, index))
+			.GroupBy(x => (x.source.Url ?? x.source.FilePath)?.ToLowerInvariant())
+			.Where(g => g.Key is not null && g.Count() > 1);
+
+		foreach (var group in sourceGroups)
+		{
+			var indices = string.Join(", ", group.Select(x => x.index));
+			yield return new ValidationResult(
+				string.Format(CoreConstants.DuplicateSourceFound, group.Key, indices),
+				[nameof(Sources)]);
+		}
+
 		// Check for duplicate VirtualPrefix values
 		var virtualPrefixGroups = Sources
 			.Select((source, index) => (source, index))
@@ -146,26 +174,46 @@ public class KoalesceOptions : IValidatableObject
 		for (int i = 0; i < Sources.Count; i++)
 		{
 			var source = Sources[i];
+			bool hasUrl = !string.IsNullOrWhiteSpace(source.Url);
+			bool hasFilePath = !string.IsNullOrWhiteSpace(source.FilePath);
 
-			if (string.IsNullOrWhiteSpace(source.Url))
+			// Validate Url XOR FilePath
+			if (hasUrl == hasFilePath)
 			{
 				yield return new ValidationResult(
-					$"The Source URL at index {i} cannot be empty.",
-					[$"{nameof(Sources)}[{i}].Url"]);
+					string.Format(CoreConstants.SourceMustHaveUrlOrFilePath, i),
+					[$"{nameof(Sources)}[{i}]"]);
 				continue;
 			}
 
-			// Note: if Koalesce supports non-HTTP protocols in the future (e.g., amqp://, file://),
-			// this validation will need to be relaxed or moved to the specific Provider.
-			// I'm keeping HTTP/HTTPS for now, as the Core uses HttpClient for now.
-			bool isUriValid = Uri.TryCreate(source.Url, UriKind.Absolute, out Uri? uriResult)
-				&& (uriResult.Scheme == Uri.UriSchemeHttp || uriResult.Scheme == Uri.UriSchemeHttps);
-
-			if (!isUriValid)
+			// Validate URL format
+			if (hasUrl)
 			{
-				yield return new ValidationResult(
-					$"The Source URL at index {i} ('{source.Url}') must be a valid absolute URL (starting with http:// or https://).",
-					[$"{nameof(Sources)}[{i}].Url"]);
+				bool isUriValid = Uri.TryCreate(source.Url, UriKind.Absolute, out Uri? uriResult)
+					&& (uriResult.Scheme == Uri.UriSchemeHttp || uriResult.Scheme == Uri.UriSchemeHttps);
+
+				if (!isUriValid)
+				{
+					yield return new ValidationResult(
+						$"The Source URL at index {i} ('{source.Url}') must be a valid absolute URL (starting with http:// or https://).",
+						[$"{nameof(Sources)}[{i}].Url"]);
+				}
+			}
+
+			// Validate FilePath exists
+			if (hasFilePath)
+			{
+				// Resolve relative paths against the application base directory
+				string resolvedPath = Path.IsPathRooted(source.FilePath!)
+					? source.FilePath!
+					: Path.Combine(AppContext.BaseDirectory, source.FilePath!);
+
+				if (!File.Exists(resolvedPath))
+				{
+					yield return new ValidationResult(
+						string.Format(CoreConstants.SourceFilePathNotFound, i, source.FilePath),
+						[$"{nameof(Sources)}[{i}].FilePath"]);
+				}
 			}
 
 			// Validate ExcludePaths
