@@ -14,11 +14,7 @@ global using Microsoft.Extensions.DependencyInjection.Extensions;
 global using Microsoft.Extensions.Logging;
 global using Microsoft.Extensions.Options;
 global using Microsoft.OpenApi;
-global using Microsoft.OpenApi.Exceptions;
-global using Microsoft.OpenApi.Extensions;
-global using Microsoft.OpenApi.Models;
-global using Microsoft.OpenApi.Readers;
-global using Microsoft.OpenApi.Services;
+global using Microsoft.OpenApi.Reader;
 global using System;
 global using System.Collections.Generic;
 global using System.ComponentModel.DataAnnotations;
