@@ -1,44 +1,227 @@
-﻿namespace Koalesce.OpenAPI.Services.ConflictResolution;
+namespace Koalesce.OpenAPI.Services.ConflictResolution;
 
 /// <summary>
-/// Internal visitor to rewrite Schema reference Ids in the source document
-/// Used when renaming schemas to avoid conflicts
+/// Service responsible for rewriting Schema reference Ids in OpenAPI documents.
 /// </summary>
-internal class SchemaReferenceRewriter : OpenApiVisitorBase
+internal static class SchemaReferenceRewriter
 {
-	private readonly IReadOnlyDictionary<string, string> _renames;
-
-	public SchemaReferenceRewriter(IReadOnlyDictionary<string, string> renames)
+	/// <summary>
+	/// Rewrites all schema references in the document according to the rename map.
+	/// </summary>
+	public static void RewriteReferences(OpenApiDocument document, IReadOnlyDictionary<string, string> renames)
 	{
-		_renames = renames;
+		if (renames.Count == 0) 
+			return;
+
+		if (document.Paths is null) 
+			return;
+
+		foreach (var pathItem in document.Paths.Values)
+		{
+			if (pathItem.Operations is null) 
+				continue;
+
+			foreach (var operation in pathItem.Operations.Values)
+			{
+				// Rewrite request body schema references
+				RewriteRequestBody(operation.RequestBody, renames, document);
+
+				// Rewrite response schema references
+				if (operation.Responses is not null)
+				{
+					foreach (var response in operation.Responses.Values)
+						RewriteResponse(response, renames, document);
+				}
+
+				// Rewrite parameter schema references
+				RewriteParameters(operation.Parameters, renames, document);
+			}
+
+			// Rewrite path-level parameter schema references
+			RewriteParameters(pathItem.Parameters, renames, document);
+		}
+
+		// Rewrite schema references within component schemas (nested refs)
+		if (document.Components?.Schemas is not null)
+		{
+			foreach (var kvp in document.Components.Schemas.ToList())
+			{
+				var rewrittenSchema = RewriteSchemaDeep(kvp.Value, renames, document);
+				if (!ReferenceEquals(rewrittenSchema, kvp.Value))
+					document.Components.Schemas[kvp.Key] = rewrittenSchema;
+			}
+		}
 	}
 
-	public override void Visit(OpenApiSchema schema)
+	/// <summary>
+	/// Rewrites the schemas within the content of the specified OpenAPI request body according to the provided property
+	/// renaming map.
+	/// </summary>	
+	private static void RewriteRequestBody(
+		IOpenApiRequestBody? requestBody, 
+		IReadOnlyDictionary<string, string> renames, 
+		OpenApiDocument document)
 	{
-		if (schema?.Reference is null)
+		if (requestBody?.Content is null) 
 			return;
 
-		// Only rewrite component schema references
-		if (schema.Reference.Type != ReferenceType.Schema)
-			return;
-
-		if (!_renames.TryGetValue(schema.Reference.Id, out var newId))
-			return;
-
-		// OpenApiReference.ReferenceV3 is read-only (computed)
-		// Replacing the entire reference object with a new Id
-		var oldRef = schema.Reference;
-
-		schema.Reference = new OpenApiReference
+		foreach (var kvp in requestBody.Content.ToList())
 		{
-			Type = ReferenceType.Schema,
-			Id = newId,
+			var mediaType = kvp.Value;
+			if (mediaType.Schema is null) 
+				continue;
 
-			// Preserve external references if they exist (rare but possible)
-			ExternalResource = oldRef.ExternalResource
-		};
+			// Use RewriteSchemaDeep to handle nested references
+			var newSchema = RewriteSchemaDeep(mediaType.Schema, renames, document);
+			if (!ReferenceEquals(newSchema, mediaType.Schema))
+				mediaType.Schema = newSchema;
+		}
+	}
 
-		// Ensure the schema remains a reference-only schema (no inline leftovers)
-		schema.UnresolvedReference = false;
+	/// <summary>
+	/// Rewrites the schemas in the response content to reflect renamed components according to the provided mapping.
+	/// </summary>	
+	private static void RewriteResponse(
+		IOpenApiResponse response, 
+		IReadOnlyDictionary<string, string> renames, 
+		OpenApiDocument document)
+	{
+		if (response.Content is null) 
+			return;
+
+		foreach (var kvp in response.Content.ToList())
+		{
+			var mediaType = kvp.Value;
+			if (mediaType.Schema is null) 
+				continue;
+
+			// Use RewriteSchemaDeep to handle nested references
+			var newSchema = RewriteSchemaDeep(mediaType.Schema, renames, document);
+			if (!ReferenceEquals(newSchema, mediaType.Schema))
+				mediaType.Schema = newSchema;
+		}
+	}
+
+	/// <summary>
+	/// Rewrites the schemas of the specified OpenAPI parameters by applying the provided renaming rules.
+	/// </summary>	
+	private static void RewriteParameters(
+		IList<IOpenApiParameter>? parameters, 
+		IReadOnlyDictionary<string, string> renames, 
+		OpenApiDocument document)
+	{
+		if (parameters is null) 
+			return;
+
+		foreach (var param in parameters)
+		{
+			if (param is OpenApiParameter concreteParam && concreteParam.Schema is not null)
+			{
+				// Use RewriteSchemaDeep to handle nested references
+				var newSchema = RewriteSchemaDeep(concreteParam.Schema, renames, document);
+				if (!ReferenceEquals(newSchema, concreteParam.Schema))
+					concreteParam.Schema = newSchema;
+			}
+		}
+	}
+
+	/// <summary>
+	/// Recursively rewrites schema references, returning a potentially new schema.
+	/// </summary>
+	private static IOpenApiSchema RewriteSchemaDeep(
+		IOpenApiSchema schema,
+		IReadOnlyDictionary<string, string> renames,
+		OpenApiDocument document)
+	{
+		// If it's a reference that needs renaming, create a new reference
+		if (schema is OpenApiSchemaReference schemaRef)
+		{
+			var newRef = TryCreateRenamedReference(schemaRef, renames);
+			return newRef ?? schema;
+		}
+
+		// If it's a concrete schema, recursively process nested schemas
+		if (schema is OpenApiSchema concreteSchema)
+		{
+			// Items (for arrays)
+			if (concreteSchema.Items is not null)
+			{
+				var newItems = RewriteSchemaDeep(concreteSchema.Items, renames, document);
+				if (!ReferenceEquals(newItems, concreteSchema.Items))
+					concreteSchema.Items = newItems;
+			}
+
+			// Properties (for objects)
+			if (concreteSchema.Properties is not null)
+			{
+				foreach (var kvp in concreteSchema.Properties.ToList())
+				{
+					var newProp = RewriteSchemaDeep(kvp.Value, renames, document);
+					if (!ReferenceEquals(newProp, kvp.Value))
+						concreteSchema.Properties[kvp.Key] = newProp;
+				}
+			}
+
+			// AllOf
+			RewriteSchemaList(concreteSchema.AllOf, renames, document);
+
+			// OneOf
+			RewriteSchemaList(concreteSchema.OneOf, renames, document);
+
+			// AnyOf
+			RewriteSchemaList(concreteSchema.AnyOf, renames, document);
+
+			// AdditionalProperties
+			if (concreteSchema.AdditionalProperties is IOpenApiSchema additionalSchema)
+			{
+				var newAdditional = RewriteSchemaDeep(additionalSchema, renames, document);
+				if (!ReferenceEquals(newAdditional, additionalSchema))
+					concreteSchema.AdditionalProperties = newAdditional;
+			}
+		}
+
+		return schema;
+	}
+
+	/// <summary>
+	/// Rewrites each schema in the specified list by applying the provided renaming rules to schema references within the
+	/// context of the given OpenAPI document.
+	/// </summary>	
+	private static void RewriteSchemaList(
+		IList<IOpenApiSchema>? schemas,
+		IReadOnlyDictionary<string, string> renames,
+		OpenApiDocument document)
+	{
+		if (schemas is null) 
+			return;
+
+		for (int i = 0; i < schemas.Count; i++)
+		{
+			var newSchema = RewriteSchemaDeep(schemas[i], renames, document);
+			if (!ReferenceEquals(newSchema, schemas[i]))
+				schemas[i] = newSchema;
+		}
+	}
+
+	/// <summary>
+	/// Creates a new OpenApiSchemaReference with the renamed Id if the schema needs renaming.
+	/// Returns null if no renaming is needed.
+	/// </summary>
+	private static OpenApiSchemaReference? TryCreateRenamedReference(
+		IOpenApiSchema schema,
+		IReadOnlyDictionary<string, string> renames)
+	{
+		if (schema is not OpenApiSchemaReference schemaRef)
+			return null;
+
+		var reference = schemaRef.Reference;
+		if (reference?.Id is null)
+			return null;
+
+		if (!renames.TryGetValue(reference.Id, out var newId))
+			return null;
+
+		// Create a new reference with the renamed Id
+		return new OpenApiSchemaReference(newId);
 	}
 }
